+++ ibis/backends/impala/__init__.py
@@ -1,15 +1,16 @@
+"""Impala backend"""
 import ibis.common.exceptions as com
 from ibis.config import options
 
 # these objects are exposed in the public API and are not used in the module
-from ibis.impala.client import (  # noqa: F401
+from .client import (  # noqa: F401
     ImpalaClient,
     ImpalaConnection,
     ImpalaDatabase,
     ImpalaTable,
 )
-from ibis.impala.compiler import dialect  # noqa: F401
-from ibis.impala.udf import *  # noqa: F401,F403
+from .compiler import dialect  # noqa: F401
+from .udf import *  # noqa: F401,F403
 
 
 def compile(expr, params=None):
