+++ ibis/backends/impala/ddl.py
