+++ ibis/backends/impala/identifiers.py
