+++ ibis/backends/impala/metadata.py
