+++ ibis/backends/impala/tests/conftest.py
