+++ ibis/backends/impala/tests/test_client.py
