+++ ibis/backends/impala/tests/test_connection_pool.py
