+++ ibis/backends/impala/tests/test_ddl.py
