+++ ibis/backends/impala/tests/test_ddl_compilation.py
