+++ ibis/backends/impala/tests/test_exprs.py
@@ -12,12 +12,12 @@ import ibis.expr.types as ir
 from ibis import literal as L
 from ibis.common.exceptions import RelationError
 from ibis.expr.datatypes import Category
-from ibis.tests.expr.mocks import MockConnection
 from ibis.impala.compiler import (  # noqa: E402
     ImpalaDialect,
     ImpalaExprTranslator,
     to_sql,
 )
+from ibis.tests.expr.mocks import MockConnection
 from ibis.tests.sql.test_compiler import ExprTestCases  # noqa: E402
 
 pytest.importorskip('hdfs')
