+++ ibis/backends/impala/tests/test_kudu_support.py
