+++ ibis/backends/impala/tests/test_metadata.py
