+++ ibis/backends/impala/tests/test_pandas_interop.py
