+++ ibis/backends/impala/tests/test_partition.py
