+++ ibis/backends/impala/tests/test_patched.py
