+++ ibis/backends/impala/tests/test_sql.py
